requirements.txt
attrs==16.2.0
beautifulsoup4==4.5.1
cffi==1.8.3
click==6.6
cryptography==1.5.2
cssselect==0.9.2
decorator==4.0.10
html2text==2016.9.19
idna==2.1
lxml==3.6.4
Markdown==2.6.7
parsel==1.0.3
praw==4.0.0b20
prawcore==0.2.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
pyOpenSSL==16.1.0
python-dotenv==0.6.0
queuelib==1.4.2
requests==2.11.1
service-identity==16.0.0
six==1.10.0
Twisted==16.4.1
update-checker==0.12
w3lib==1.15.0
zope.interface==4.3.2
